net/wireguard - offer CARP vhid tracking support, closes https://github.com/opnsense/plugins/issues/3579
When the the selected vhid is in BACKUP or INIT mode, the wireguard interface in question will be set to "down", in which case communication stops and the new master may take over. The advantage of this strategy is that switching is relatively quick as only the interface flag need to be changed.